+++ src/main/java/com/tteokguk/tteokguk/item/exception/ItemError.java
@@ -1,0 +1,21 @@
+package com.tteokguk.tteokguk.item.exception;
+
+import com.tteokguk.tteokguk.global.exception.error.ErrorCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
+
+@Getter
+@RequiredArgsConstructor
+public enum ItemError implements ErrorCode {
+
+    INSUFFICIENT_INGREDIENTS("해당 사용자는 재료가 부족합니다.", BAD_REQUEST, "I_001"),
+    INGREDIENTS_COUNT_CONFLICT("재료는 5개만 요청 가능합니다", BAD_REQUEST, "I_002"),
+    ;
+
+    private final String message;
+    private final HttpStatus status;
+    private final String code;
+}
