+++ src/main/java/com/tteokguk/tteokguk/item/infra/persistence/ItemRepository.java
@@ -1,9 +1,8 @@
-package com.tteokguk.tteokguk.member.infra.persistence;
+package com.tteokguk.tteokguk.item.infra.persistence;
 
+import com.tteokguk.tteokguk.item.domain.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.tteokguk.tteokguk.member.domain.Item;
 
 @Transactional
 public interface ItemRepository extends JpaRepository<Item, Long> {
