+++ src/main/java/com/tteokguk/tteokguk/member/application/dto/response/MyPageResponse.java
@@ -1,5 +1,7 @@
 package com.tteokguk.tteokguk.member.application.dto.response;
 
+import com.tteokguk.tteokguk.item.application.dto.response.ItemResponse;
+import com.tteokguk.tteokguk.tteokguk.application.dto.response.SimpleTteokgukResponse;
 import com.tteokguk.tteokguk.tteokguk.constants.Ingredient;
 import lombok.Builder;
 
@@ -10,7 +12,7 @@ public record MyPageResponse(
         Long id,
         Ingredient primaryIngredient,
         String nickname,
-        List<TteokgukResponse> tteokguks,
+        List<SimpleTteokgukResponse> tteokguks,
         List<ItemResponse> items
 ) {
 }
