+++ src/main/java/com/tteokguk/tteokguk/tteokguk/application/dto/response/TteokgukResponse.java
@@ -1,4 +1,4 @@
-package com.tteokguk.tteokguk.member.application.dto.response;
+package com.tteokguk.tteokguk.tteokguk.application.dto.response;
 
 import com.tteokguk.tteokguk.tteokguk.constants.Ingredient;
 import lombok.Builder;
@@ -8,6 +8,7 @@ import java.util.List;
 @Builder
 public record TteokgukResponse(
         Long tteokgukId,
+        Long memberId,
         String wish,
         boolean access,
         boolean completion,
