+++ src/main/java/com/tteokguk/tteokguk/tteokguk/application/dto/response/assembler/TteokgukResponseAssembler.java
@@ -1,6 +1,6 @@
 package com.tteokguk.tteokguk.tteokguk.application.dto.response.assembler;
 
-import com.tteokguk.tteokguk.tteokguk.application.dto.response.CreateTteokgukResponse;
+import com.tteokguk.tteokguk.tteokguk.application.dto.response.TteokgukResponse;
 import com.tteokguk.tteokguk.tteokguk.domain.Tteokguk;
 import lombok.NoArgsConstructor;
 
@@ -8,8 +8,8 @@ import static lombok.AccessLevel.PRIVATE;
 
 @NoArgsConstructor(access = PRIVATE)
 public class TteokgukResponseAssembler {
-    public static CreateTteokgukResponse toTteokgukResponse(Tteokguk tteokguk) {
-        return CreateTteokgukResponse.builder()
+    public static TteokgukResponse toTteokgukResponse(Tteokguk tteokguk) {
+        return TteokgukResponse.builder()
                 .tteokgukId(tteokguk.getId())
                 .memberId(tteokguk.getMember().getId())
                 .wish(tteokguk.getWish())
